requirements.txt
Flask
pymongo
python-dotenv
cloudinary
openai-whisper
transformers
torch 
flask-cors 